Refactoring Some of the scripts.
A bit of house cleanup before getting into the splines.
In specific, the curve manager will now take care of all the curves, in terms of their active state, visibility etc, and a curve controller, added to a curve group (where the spline will also be added) will be controlling the specific curve tasks, like adding and managing marks.