Curve Mesh
I managed to create the curve mesh, I needed a mesh to detect raycast collision with the mouse, so the next step is to be able to detect collision with the curve, and use that to say that the player wants to add a marker inbetween markers rather than ouside.
I'm having an issue that I don't know how to fix though. Since I'm calculating the rotation both of the markers and the mesh points with the tangents and upVectors of the knots, and they rotate very abrubtly in certain situations, it makes my mesh get thin when certain rotations happen, because the upvector is flipping 180 degrees. I don't know how to fix this. But for now, I'll keep this as it is.
I probably won't render the mesh, rather I'll only use it for collision detection, and will render the mesh with Shapes package.