Frame Numbers
Besides showing the frame numbers, I also started with the UI that will accept frame changes, but didn't manage to connect that to the functionality. Need to think this through.
I think my struggle has been that the selection logic is in Curve Controller and the connections are very messy.